--- bsp/inc/BSP_CAN.h
@@ -1,8 +1,0 @@
-#ifndef __BSP_CAN_H
-#define __BSP_CAN_H
-
-#include "stm32f4xx.h"
-
-void BSP_CAN_Init(void);
-
-#endif
