--- bsp/inc/BSP_DMA.h
@@ -1,8 +1,0 @@
-#ifndef __BSP_DMA_H
-#define __BSP_DMA_H
-
-#include "stm32f4xx.h"
-
-void BSP_DMA_Init(void);
-
-#endif
