--- bsp/inc/BSP_GPIO.h
@@ -1,8 +1,0 @@
-#ifndef __BSP_GPIO_H
-#define __BSP_GPIO_H
-
-#include "stm32f4xx.h"
-
-void BSP_GPIO_Init(void);
-
-#endif
