--- bsp/inc/BSP_NVIC.h
@@ -1,8 +1,0 @@
-#ifndef __BSP_NVIC_H
-#define __BSP_NVIC_H
-
-#include "stm32f4xx.h"
-
-void BSP_NVIC_Init(void);
-
-#endif
