--- bsp/inc/BSP_TIM.h
@@ -1,8 +1,0 @@
-#ifndef __BSP_TIM_H
-#define __BSP_TIM_H
-
-#include "stm32f4xx.h"
-
-void BSP_TIM_Init(void);
-
-#endif
