--- bsp/inc/BSP_UART.h
@@ -1,8 +1,0 @@
-#ifndef __BSP_UART_H
-#define __BSP_UART_H
-
-#include "stm32f4xx.h"
-
-void BSP_UART_Init(void);
-
-#endif
