--- bsp/src/BSP_DMA.c
@@ -1,34 +1,0 @@
-#include "BSP_DMA.h"
-#include "config.h"
-#include "handle.h"
-
-/**
- * @brief  DMA初始化
- * @param  void
- * @return void
- */
-void BSP_DMA_Init(void) {
-    DMA_InitTypeDef DMA_InitStructure;
-
-    //	RCC_AHB1PeriphClockCmd(RCC_AHB1Periph_DMA1, ENABLE);     no use currently
-    RCC_AHB1PeriphClockCmd(RCC_AHB1Periph_DMA2, ENABLE);
-
-    // UART1(DBus)
-    DMA_InitStructure.DMA_Channel            = DMA_Channel_4;
-    DMA_InitStructure.DMA_PeripheralBaseAddr = (uint32_t)(&USART1->DR);
-    DMA_InitStructure.DMA_Memory0BaseAddr    = (uint32_t)(remoteBuffer);
-    DMA_InitStructure.DMA_DIR                = DMA_DIR_PeripheralToMemory;
-    DMA_InitStructure.DMA_BufferSize         = DBUS_LENGTH + DBUS_BACK_LENGTH;
-    DMA_InitStructure.DMA_PeripheralInc      = DMA_PeripheralInc_Disable;
-    DMA_InitStructure.DMA_MemoryInc          = DMA_MemoryInc_Enable;
-    DMA_InitStructure.DMA_MemoryDataSize     = DMA_MemoryDataSize_Byte;
-    DMA_InitStructure.DMA_PeripheralDataSize = DMA_PeripheralDataSize_Byte;
-    DMA_InitStructure.DMA_Mode               = DMA_Mode_Normal;
-    DMA_InitStructure.DMA_Priority           = DMA_Priority_Medium;
-    DMA_InitStructure.DMA_FIFOMode           = DMA_FIFOMode_Disable;
-    DMA_InitStructure.DMA_FIFOThreshold      = DMA_FIFOThreshold_Full;
-    DMA_InitStructure.DMA_MemoryBurst        = DMA_MemoryBurst_Single;
-    DMA_InitStructure.DMA_PeripheralBurst    = DMA_PeripheralBurst_Single;
-    DMA_Init(DMA2_Stream2, &DMA_InitStructure);
-    DMA_Cmd(DMA2_Stream2, ENABLE);
-}
