--- bsp/src/BSP_NVIC.c
@@ -1,80 +1,0 @@
-#include "BSP_NVIC.h"
-#include "config.h"
-
-/**
- * @brief  中断初始化
- * @param  void
- * @return void
- * USART6 用作无线串口,定义在usart.c中
- */
-void BSP_NVIC_Init(void) {
-    NVIC_InitTypeDef NVIC_InitStructure;
-    EXTI_InitTypeDef EXTI_InitStructure;
-
-    // UART1(DBus)
-    NVIC_InitStructure.NVIC_IRQChannel                   = USART1_IRQn;
-    NVIC_InitStructure.NVIC_IRQChannelCmd                = ENABLE;
-    NVIC_InitStructure.NVIC_IRQChannelPreemptionPriority = 8;
-    NVIC_InitStructure.NVIC_IRQChannelSubPriority        = 0;
-    NVIC_Init(&NVIC_InitStructure);
-    USART_ITConfig(USART1, USART_IT_IDLE, ENABLE);
-
-    // USART3
-#if USART3_ENABLED
-    NVIC_InitStructure.NVIC_IRQChannel                   = USART3_IRQn; //串口3中断通道
-    NVIC_InitStructure.NVIC_IRQChannelPreemptionPriority = 7;           //抢占优先级3
-    NVIC_InitStructure.NVIC_IRQChannelSubPriority        = 0;           //子优先级3
-    NVIC_InitStructure.NVIC_IRQChannelCmd                = ENABLE;      // IRQ通道使能
-    NVIC_Init(&NVIC_InitStructure);                                     //根据指定的参数初始化VIC寄存器
-#endif
-
-    // USART6
-#if USART6_ENABLED
-    NVIC_InitStructure.NVIC_IRQChannel                   = USART6_IRQn; //串口3中断通道
-    NVIC_InitStructure.NVIC_IRQChannelPreemptionPriority = 7;           //抢占优先级3
-    NVIC_InitStructure.NVIC_IRQChannelSubPriority        = 0;           //子优先级3
-    NVIC_InitStructure.NVIC_IRQChannelCmd                = ENABLE;      // IRQ通道使能
-    NVIC_Init(&NVIC_InitStructure);                                     //根据指定的参数初始化VIC寄存器
-#endif
-
-    // Gyroscope (NVIC)
-    NVIC_InitStructure.NVIC_IRQChannel                   = EXTI9_5_IRQn;
-    NVIC_InitStructure.NVIC_IRQChannelPreemptionPriority = 8;
-    NVIC_InitStructure.NVIC_IRQChannelSubPriority        = 0;
-    NVIC_InitStructure.NVIC_IRQChannelCmd                = ENABLE;
-    NVIC_Init(&NVIC_InitStructure);
-
-    // Gyroscope (EXTI)
-    SYSCFG_EXTILineConfig(EXTI_PortSourceGPIOB, GPIO_PinSource8);
-    EXTI_InitStructure.EXTI_Line    = EXTI_Line8;
-    EXTI_InitStructure.EXTI_Mode    = EXTI_Mode_Interrupt;
-    EXTI_InitStructure.EXTI_Trigger = EXTI_Trigger_Falling; //下降沿中断
-    EXTI_InitStructure.EXTI_LineCmd = ENABLE;
-    EXTI_Init(&EXTI_InitStructure);
-
-    // CAN1
-    NVIC_InitStructure.NVIC_IRQChannel                   = CAN1_RX0_IRQn;
-    NVIC_InitStructure.NVIC_IRQChannelCmd                = ENABLE;
-    NVIC_InitStructure.NVIC_IRQChannelPreemptionPriority = 8;
-    NVIC_InitStructure.NVIC_IRQChannelSubPriority        = 0;
-    NVIC_Init(&NVIC_InitStructure);
-    CAN_ITConfig(CAN1, CAN_IT_EWG | CAN_IT_EPV | CAN_IT_BOF | CAN_IT_LEC | CAN_IT_ERR | CAN_IT_WKU | CAN_IT_SLK, ENABLE);
-    CAN1->IER |= 0xFF;
-
-    // CAN2
-    NVIC_InitStructure.NVIC_IRQChannel                   = CAN2_RX0_IRQn;
-    NVIC_InitStructure.NVIC_IRQChannelCmd                = ENABLE;
-    NVIC_InitStructure.NVIC_IRQChannelPreemptionPriority = 8;
-    NVIC_InitStructure.NVIC_IRQChannelSubPriority        = 0;
-    NVIC_Init(&NVIC_InitStructure);
-    CAN_ITConfig(CAN2, CAN_IT_FMP0, ENABLE);
-
-    // TIM2
-    NVIC_InitStructure.NVIC_IRQChannel                   = TIM2_IRQn;
-    NVIC_InitStructure.NVIC_IRQChannelPreemptionPriority = 9;
-    NVIC_InitStructure.NVIC_IRQChannelSubPriority        = 0;
-    NVIC_InitStructure.NVIC_IRQChannelCmd                = ENABLE;
-    NVIC_Init(&NVIC_InitStructure);
-    TIM_ITConfig(TIM2, TIM_IT_Update, ENABLE);
-    TIM_ClearFlag(TIM2, TIM_FLAG_Update);
-}
