+++ user/template/bsp.c
@@ -1,10 +1,15 @@
 /**
- * @brief 用户BSP
+ * @brief 硬件初始化
  * @note 该函数应该在task.c中被执行
  */
 
 #include "bsp.h"
+#include "handle.h"
+#include "Driver_BSP.h"
 
-void BSP_USER_Init(void) {
-    int do_whatever_you_like_here = 0;
+void BSP_Init(void) {
+    BSP_CAN_Init();
+    BSP_DBUS_Init(remoteBuffer);
+    BSP_TIM2_Init();
+    BSP_USART6_Init(115200);
 }
